+++ motion_control/metas/polar_parallel_meta_controller/include/polar_parallel_meta_controller/PolarParallelMetaController.hpp
@@ -73,9 +73,9 @@ protected:
         // Angular command
         outputs_[1] = angular_ctrl->output(0);
         // Pose reached has been set by previous filter except if blocked
-        if ((linear_ctrl->output(1) == (double)target_pose_status_t::blocked)
-            || (angular_ctrl->output(1) == (double)target_pose_status_t::blocked)) {
-            outputs_[2] = (double)target_pose_status_t::blocked;
+        if ((linear_ctrl->output(1) == (float)target_pose_status_t::blocked)
+            || (angular_ctrl->output(1) == (float)target_pose_status_t::blocked)) {
+            outputs_[2] = (float)target_pose_status_t::blocked;
         }
         else {
             outputs_[2] = inputs_[8];
